requirements.txt
fastapi==0.115.12
uvicorn[standard]==0.34.0
redis==5.2.1
rq==2.1.0
pydantic==2.10.6
pydantic-settings==2.7.1
python-multipart==0.0.20
pytest==8.3.4
pytest-asyncio==0.25.3
httpx==0.28.1
pytesseract==0.3.13
Pillow==11.1.0